Refactor: naver,gmail관련 코드 리팩토링 및 통합
