--- computerNetworkAssignment/src/WebSocketApplication.java
@@ -1,8 +1,0 @@
-import java.io.IOException;
-
-public class WebSocketApplication {
-    public static void main(String[] args) throws IOException {
-        UI webSocketProgram = new UI();
-        webSocketProgram.run();
-    }
-}
