+++ webSocket/src/main/java/computerNetwork/webSocket/WebSocketApplication.java
@@ -4,15 +4,14 @@ package computerNetwork.webSocket;
 
 
 
-//import computerNetwork.webSocket.gmail.GmailFetcher;
 
-import computerNetwork.webSocket.gmail.GmailMailbox;
+import computerNetwork.webSocket.gmail.GmailFetcher;
 
 import java.io.IOException;
 
 public class WebSocketApplication {
 	public static void main(String[] args) throws IOException {
-	GmailMailbox gmailFetcher=new GmailMailbox();
+	GmailFetcher gmailFetcher=new GmailFetcher();
 	gmailFetcher.checkEmails("mhg10181018@gmail.com","dkgoanypxohpizoi");
 	}
